2018 On the origin of the Epps Effect
Nice literature review and complete introduction. In this paper, an analytic decomposition of cross-correlation function of asynchronous events using time lagged correlations is made after studying the evolution of the characteristic time. The characteristic time tends to decrease as trading activity increase. In the last part, they propose a model of this decomposition using random walks with asynchronous steps.