1993 Cross-autocorrelation stock prices
In this paper, a model is created to study to cross-autocorrelation among stocks. This also explain why correlations are positive and how big firms drive lower firms. The model supports the idea that nonsynchronous trading implies an Epps effect. The idea of the model is that market makers observe noisy signals but can't adjust prices instantly with respect to other stocks signals that also contain market informations.