2018 On the origin of the Epps Effect
The aim of the present communication is to investigate empirically the role of the trading activity on the Epps effect. We find that in contrast to the simulations, the empirical data do not scale in the expected way indicating that a further, possibly human factor is also at play. The possible human factor is deduced by the fact that characteristic time is independent of trading frequency which is quite surprising.