2007 Modeling the Epps effect ...
In this paper, a microstructure of lagged cross-correlations among returns using both approaches: a distributed lag model that includes lagged trade and quote data for both assets, as well as, a modified or pseudo-error correction mechanism. This is a way to modelize the  Chan (1993) model where market maker adjust price with a lag. The approach is new in this paper because doesn't use lead–lag or error correction models (most models in the literature). These lead-lag and correction models are well described in part 2.